auth/service/requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pika==1.3.2
psycopg2==2.9.9
PyJWT==2.9.0
pyotp==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
